refactor: Clothes Controller Command, Query분리
